# specs/005-core-domain-crud-apis/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AkarAI Phase 4 Core Domain CRUD API
  version: 0.1.0
  description: >
    Contract for Phase 4 core domain database and CRUD APIs. This contract excludes
    AI search, RAG, image upload/processing, OCR, email sending, dashboards,
    chatbot behavior, buyer-to-agency real-time chat, spam classification,
    lead scoring, and generated replies.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
paths:
...
  /agency/listings:
    get:
      summary: List tenant listings
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paginated tenant listings }
    post:
      summary: Create listing
      tags: [Agency Listings]
      responses:
        "201": { description: Listing created }
        "403": { description: Support employees cannot create listings }
  /agency/listings/{listing_id}:
    get:
      summary: Get tenant listing
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200": { description: Listing }
    patch:
      summary: Update tenant listing
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200": { description: Listing updated }
    delete:
      summary: Archive tenant listing
      tags: [Agency Listings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "204": { description: Listing archived }
...
  /agency/listings/{listing_id}/viewing-slots:
    get:
      summary: List listing viewing slots
      tags: [Viewing Slots]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200": { description: Viewing slots }
    post:
      summary: Create listing viewing slot
      tags: [Viewing Slots]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "201": { description: Viewing slot created }
        "403": { description: Support employees cannot create viewing slots unless explicitly granted later }
...
  /listings/{listing_id}/viewings:
    post:
      summary: Schedule listing viewing
      tags: [Viewings]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "201": { description: Scheduled viewing and initial status history created }
        "409": { description: Viewing slot unavailable }
        "429": { description: Viewing booking rate limit exceeded }
  /me/viewings:
    get:
      summary: List my scheduled viewings
      tags: [Viewings]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paginated user scheduled viewings }
  /me/viewings/{viewing_id}:
    get:
      summary: Get my scheduled viewing
      tags: [Viewings]
      parameters:
        - $ref: "#/components/parameters/ViewingId"
      responses:
        "200": { description: User scheduled viewing detail }
  /agency/viewings:
    get:
      summary: List agency scheduled viewings
      tags: [Viewings]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paginated tenant scheduled viewings }
  /agency/viewings/{viewing_id}:
    get:
      summary: Get agency scheduled viewing
      tags: [Viewings]
      parameters:
        - $ref: "#/components/parameters/ViewingId"
      responses:
        "200": { description: Tenant scheduled viewing detail }
    patch:
      summary: Update scheduled viewing status
      tags: [Viewings]
      parameters:
        - $ref: "#/components/parameters/ViewingId"
      responses:
        "200": { description: Scheduled viewing updated and status history appended }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    ListingId:
      name: listing_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ViewingId:
      name: viewing_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...